[rm] libraries

+++ 0x08-palindrome_integer/0-is_palindrome.c
@@ -7,7 +7,7 @@
  */
 int is_palindrome(unsigned long n)
 {
-	unsigned long long int big_zero = 1000000000000000000;
+	unsigned long int big_zero = 1000000000000000000;
 	int top = 0;
 	int bot = 0;
